My commit for the day

+++ sh142/sh142/jobs.cpp
@@ -7,7 +7,42 @@
 //
 
 #include "jobs.h"
-//#include "main.h"
+
+void shellInit()
+{
+    SHELL_PID = getpid();
+    SHELL_TERMINAL = STDIN_FILENO;
+    SHELL_IS_INTERACTIVE = isatty(SHELL_TERMINAL);
+    
+    if (SHELL_IS_INTERACTIVE) {
+        /*while (tcgetpgrp(SHELL_TERMINAL) != (SHELL_PGID = getpgrp())) {
+            kill(SHELL_PID, SIGCONT);
+        }*/
+        
+        signal(SIGQUIT, SIG_IGN);
+        signal(SIGTTOU, SIG_IGN);
+        signal(SIGTTIN, SIG_IGN);
+        signal(SIGTSTP, SIG_IGN);
+        signal(SIGINT, SIG_IGN);
+        signal(SIGCHLD, &childSignalHandler);
+        
+        setpgid(SHELL_PID, SHELL_PID);
+        SHELL_PGID = getpgrp();
+        if (SHELL_PID != SHELL_PGID) {
+            errormsg((char*)"Shell is not the process group leader");
+            exit(EXIT_FAILURE);
+        }
+        if (tcsetpgrp(SHELL_TERMINAL, SHELL_PGID) == -1) {
+            tcgetattr(SHELL_TERMINAL, &SHELL_TMODES);
+        }
+        
+    }
+    else {
+        errormsg((char*)"Could not make the shell interactive.");
+        exit(EXIT_FAILURE);
+    }
+}
+
 
 /**
 	launchJob
@@ -26,19 +61,19 @@ int launchJob(char* cmd[], int mode, char* path, int flag)
         errormsg((char*)"Failed to fork new process");
         return EXIT_FAILURE;
     } else if (pid == 0) { //Child only
-        signal(SIGQUIT, SIG_DFL);
-        signal(SIGTTOU, SIG_DFL);
-        signal(SIGTTIN, SIG_DFL);
-        signal(SIGTSTP, SIG_DFL);
         signal(SIGINT, SIG_DFL);
+        signal(SIGQUIT, SIG_DFL);
+        signal(SIGTSTP, SIG_DFL);
         signal(SIGCHLD, &childSignalHandler);
-        //usleep(20000);
+        signal(SIGTTIN, SIG_DFL);
+        
+        usleep(20000);
         setpgrp();
         
         if (mode == FOREGROUND) {tcsetpgrp(SHELL_TERMINAL, getpid());}
         else if (mode == BACKGROUND) {
-            printf("Job launched as process '%d'\n", getpid());
             numberOfActiveJobs++;
+            printf("Job [%d] launched as process '%d'\n",numberOfActiveJobs, getpid());
         }
         
         //Run the job
@@ -59,17 +94,20 @@ int launchJob(char* cmd[], int mode, char* path, int flag)
             return EXIT_FAILURE;
         }
         
-        exit(EXIT_SUCCESS);
-        //return EXIT_SUCCESS;
+        //exit(EXIT_SUCCESS);
+        return EXIT_SUCCESS;
     } else { //Parent only
         
         setpgid(pid, pid); //Will default to pid
         
         jobList = addJob(pid, pid, *cmd, path, mode);
         job *j = getJob(pid, PROCESSID);
+        //printf("TEST1\n");
         switch (mode) {
-            case FOREGROUND: setJobInBackground(j, 0, false); break;
-            case BACKGROUND: setJobInBackground(j, 0, true); break;
+            case FOREGROUND: jobToForeground(j, FALSE); break;
+            case BACKGROUND: jobToBackground(j, FALSE); break;
+            //case FOREGROUND: setJobInBackground(j, 0, false); break;
+            //case BACKGROUND: printf("TEST\n"); setJobInBackground(j, 0, true); break;
             default: break;
         }
         wait(NULL);
@@ -87,6 +125,7 @@ void errormsg(char* c) {
 
 job* addJob(pid_t pid, pid_t pgid, char* jobName,char* descriptor, int status)
 {
+    usleep(10000);
     job *j = (job*) malloc(sizeof(job));
     j->name = (char*) malloc(sizeof(jobName));
     j->name = strcpy(j->name, jobName);
@@ -114,76 +153,41 @@ job* addJob(pid_t pid, pid_t pgid, char* jobName,char* descriptor, int status)
     }
 }
 
-void jobInit()
-{
-    SHELL_PID = getpid();
-    SHELL_TERMINAL = STDIN_FILENO;
-    SHELL_IS_INTERACTIVE = isatty(SHELL_TERMINAL);
-    
-    if (SHELL_IS_INTERACTIVE) {
-        while (tcgetpgrp(SHELL_TERMINAL) != (SHELL_PGID = getpgrp())) {
-            kill(SHELL_PID, SIGTTIN);
-        }
-        
-        signal(SIGQUIT, SIG_IGN);
-        signal(SIGTTOU, SIG_IGN);
-        signal(SIGTTIN, SIG_IGN);
-        signal(SIGTSTP, SIG_IGN);
-        signal(SIGINT, SIG_IGN);
-        signal(SIGCHLD, &childSignalHandler);
-        
-        setpgid(SHELL_PID, SHELL_PID);
-        SHELL_PGID = getpgrp();
-        if (SHELL_PID != SHELL_PGID) {
-            errormsg((char*)"Shell is not the process group leader");
-            exit(EXIT_FAILURE);
-        }
-        if (tcsetpgrp(SHELL_TERMINAL, SHELL_PGID) == -1) {
-            tcgetattr(SHELL_TERMINAL, &SHELL_TMODES);
-        }
-        
-    }
-    else {
-        errormsg((char*)"Could not make the shell interactive.");
-        exit(EXIT_FAILURE);
-    }
-}
-
 void childSignalHandler(int i)
 {
     int status;
     pid_t pid = waitpid(WAIT_ANY, &status, WUNTRACED | WNOHANG);
     if (pid > 0) {
-        job* job = getJob(pid, PROCESSID);
-        if (job == NULL) {
+        job* j = getJob(pid, PROCESSID);
+        if (j == NULL) {
             return;
         }
         if (WIFEXITED(status)) {
-            if (job->status == BACKGROUND) {
-                printf("\nJob: ‰d Done\t ‰d\n", job->id, job->name);
-                jobList = deleteJob(job);
+            if (j->status == BACKGROUND) {
+                printf("\nJob: %d Done\t %s\n", j->id, j->name);
+                jobList = deleteJob(j);
             }
         }
         else if (WIFSIGNALED(status)) {
-            printf("\nJob: ‰d KILLED\t %s\n", job->id, job->name);
-            jobList = deleteJob(job);
+            printf("\nJob: %d KILLED\t %s\n", j->id, j->name);
+            jobList = deleteJob(j);
         }
         else if (WIFSTOPPED(status)) {
-            if (job->status == BACKGROUND) {
+            if (j->status == BACKGROUND) {
                 tcsetpgrp(SHELL_TERMINAL, SHELL_PGID);
                 setJobStatus(pid, WAITINGINPUT);
-                printf("\nJob: ‰d Suspended and wants input\t ‰s\n", numberOfActiveJobs, job->name);
+                printf("\nJob: %d Suspended and wants input\t %s\n", numberOfActiveJobs, j->name);
             }
             else {
-                tcsetpgrp(SHELL_TERMINAL, SHELL_PGID);
+                tcsetpgrp(SHELL_TERMINAL, j->pgid);
                 setJobStatus(pid, SUSPENDED);
-                printf("\nJob: %d Stopped\t %s\n", numberOfActiveJobs, job->name);
+                printf("\nJob: %d Stopped\t %s\n", numberOfActiveJobs, j->name);
             }
             return;
         }
         else {
-            if (job->status == BACKGROUND) {
-                jobList = deleteJob(job);
+            if (j->status == BACKGROUND) {
+                jobList = deleteJob(j);
             }
         }
         tcsetpgrp(SHELL_TERMINAL, SHELL_PGID);
@@ -192,19 +196,20 @@ void childSignalHandler(int i)
 
 int setJobStatus(int pid, int newStatus)
 {
-    job *job = jobList;
-    if (job == NULL) {
+    usleep(10000);
+    job *j = jobList;
+    if (j == NULL) {
         return 0;
     }
     else {
-        int counter = 0;
-        while (job != NULL) {
-            if (job->pid == pid) {
-                job->status = newStatus;
+        int i = 0;
+        while (j != NULL) {
+            if (j->pid == pid) {
+                j->status = newStatus;
                 return 1;
             }
-            counter++;
-            job = job->next;
+            i++;
+            j = j->next;
         }
         return FALSE;
     }
@@ -212,6 +217,8 @@ int setJobStatus(int pid, int newStatus)
 
 job* deleteJob(job *j)
 {
+    
+    usleep(10000);
     if (jobList == NULL) {
         return NULL;
     }
@@ -239,6 +246,7 @@ job* deleteJob(job *j)
 
 job* getJob(int value, int type)
 {
+    sleep(10000);
     job *job = jobList;
     if (type == PROCESSID) {
         while (job != NULL) {
@@ -327,6 +335,34 @@ void setJobInBackground(job* j, int cont, bool bg)
     }
 }
 
+void jobToForeground(job *j, int cont) {
+    j->status = FOREGROUND;
+    tcsetpgrp(SHELL_TERMINAL, j->pgid);
+    if (cont) {
+        if (kill(-j->pgid, SIGCONT) < 0) {
+            perror("kill (SIGCONT)");
+        }
+    }
+    waitJob(j);
+    tcsetpgrp(SHELL_TERMINAL, SHELL_PGID);
+}
+
+void jobToBackground(job *j, int cont) {
+    if (j == NULL) {
+        return;
+    }
+    if (cont && j->status == WAITINGINPUT) {
+        j->status = WAITINGINPUT;
+    }
+    if (cont) {
+        if (kill(-j->pgid, SIGCONT) < 0) {
+            perror("kill (SIGCONT)");
+        }
+    }
+    
+    tcsetpgrp(SHELL_TERMINAL, SHELL_PGID);
+}
+
 void waitJob(job* j)
 {
     int status;
@@ -341,6 +377,9 @@ void waitJob(job* j)
 void killJob(int id)
 {
     job *j = getJob(id, JOBID);
+    if (id == NULL || id > numberOfActiveJobs) {
+        return;
+    }
     kill(j->pid, SIGKILL);
 }
